updates commonality
