+++ src/controller/handler/connections.ts
@@ -9,9 +9,10 @@ import { CacheStore } from '@nestjs/common';
 /*
     Acapy webhooks handler for input received from the url [webhookurl]/v1/controller/topic/connections
 */
-export class Connections implements BaseAgentResponseHandler {
+export class Connections extends BaseAgentResponseHandler {
     private static CONNECTIONS_URL: string = 'connections';
     constructor(private readonly agentGovernance: AgentGovernance, private readonly http: ProtocolHttpService, private readonly cache: CacheStore) {
+        super();
     }
 
     private async checkPolicyForAction(governanceKey: string, cacheKey: string) {
@@ -77,16 +78,7 @@ export class Connections implements BaseAgentResponseHandler {
             await readPermission(action, templatedCacheKey);
 
             const url: string = agentUrl + `/${Connections.CONNECTIONS_URL}/${body.connection_id}/${action}`;
-            const req: AxiosRequestConfig = {
-                method: 'POST',
-                url,
-                headers: {
-                    'x-api-key': adminApiKey,
-                }
-            };
-            if (token) {
-                req.headers.Authorization = 'Bearer ' + token;
-            }
+            const req: AxiosRequestConfig = super.createHttpRequest(url, adminApiKey, token);
 
             Logger.info(`requesting agent to accept connection invite ${req.url}`);
             const res = await this.http.requestWithRetry(req);
@@ -101,16 +93,7 @@ export class Connections implements BaseAgentResponseHandler {
             await readPermission(action, templatedCacheKey);
 
             const url: string = agentUrl + `/${Connections.CONNECTIONS_URL}/${body.connection_id}/${action}`;
-            const req: AxiosRequestConfig = {
-                method: 'POST',
-                url,
-                headers: {
-                    'x-api-key': adminApiKey,
-                }
-            };
-            if (token) {
-                req.headers.Authorization = 'Bearer ' + token;
-            }
+            const req: AxiosRequestConfig = super.createHttpRequest(url, adminApiKey, token);
 
             Logger.info(`requesting initiating agent to complete connection invite ${req.url}`);
             const res = await this.http.requestWithRetry(req);
