+++ src/controller/handler/problem.report.ts
@@ -8,8 +8,9 @@ import { AgentGovernance } from '../agent.governance';
     Allows an agent to report a problem back to aries-guardianship-agency so that we
     can log it
 */
-export class ProblemReport implements BaseAgentResponseHandler {
+export class ProblemReport extends BaseAgentResponseHandler {
     constructor(private readonly agentGovernance: AgentGovernance, private readonly http: ProtocolHttpService, private readonly cache: CacheStore) {
+        super();
     }
 
     public async handlePost(
