+++ src/controller/handler/proof.ts
@@ -6,10 +6,11 @@ import { Logger } from 'protocol-common/logger';
 import { AgentGovernance } from '../agent.governance';
 import { BaseAgentResponseHandler } from './base.agent.response.handler';
 
-export class Proofs implements BaseAgentResponseHandler {
+export class Proofs extends BaseAgentResponseHandler {
     private static PROOFS_URL: string = 'present-proof';
 
     constructor(private readonly agentGovernance: AgentGovernance, private readonly http: ProtocolHttpService, private readonly cache: CacheStore) {
+        super();
     }
 
     private async checkPolicyForAction(governanceKey: string, cacheKey: string) {
@@ -136,16 +137,7 @@ export class Proofs implements BaseAgentResponseHandler {
             await readPermission(action, templatedCacheKey);
 
             const url: string = agentUrl + `/${Proofs.PROOFS_URL}/records/${body.presentation_exchange_id}/${action}`;
-            const req: AxiosRequestConfig = {
-                method: 'POST',
-                url,
-                headers: {
-                    'x-api-key': adminApiKey,
-                }
-            };
-            if (token) {
-                req.headers.Authorization = 'Bearer ' + token;
-            }
+            const req: AxiosRequestConfig = super.createHttpRequest(url, adminApiKey, token);
             Logger.info(`requesting holder to present proof ${req.url}`);
             const res = await this.http.requestWithRetry(req);
             return res.data;
@@ -186,17 +178,8 @@ export class Proofs implements BaseAgentResponseHandler {
 
             const reply = { trace: false, requested_predicates, requested_attributes, self_attested_attributes };
             url = agentUrl + `/present-proof/records/${body.presentation_exchange_id}/${action}`;
-            const req: AxiosRequestConfig = {
-                method: 'POST',
-                url,
-                headers: {
-                    'x-api-key': adminApiKey,
-                },
-                data: reply
-            };
-            if (token) {
-                req.headers.Authorization = 'Bearer ' + token;
-            }
+            const req: AxiosRequestConfig = super.createHttpRequest(url, adminApiKey, token);
+            req.data = reply;
             const res = await this.http.requestWithRetry(req);
             return res.data;
         }
